The Action function for the showing the Inspector Panel was moved into a custom window controller for the document window to allow the "Show Inspector" segue to be triggered. The action function for hiding the inspector panel was also moved here simply for consistency reasons.
A custom window controller for the inspector was also created in the hope that a singleton instance could be accessed and the hide action could easily call performClose on the window. This appears to work, but showing and hiding the inspector does not appear to handle autosaving the location of the inspector window correctly. It shows up in two different, but consistent locations.